fix: infinite loader on buy screen if stacks account is not present
